whoisspeaking/requirements.txt
uvicorn
fastapi
python-multipart
celery
asyncio
aiohttp
groq
redis
hashlib
python-dotenv
poetry